Changed create-node templates to use sh, not bash
The previous templates do not work on FreeBSD. FreeBSD does not have
bash in its default install, nor when installed is it located in /bin/
(it gets tossed in /usr/local/bin/). The shell scripts use no features
that /bin/sh does not have, so I changed the #! to /bin/sh instead,
found on FreeBSD, Mac OS X, and most flavors of Linux. The option passed
to ps was changed to display the same output on GNU and BSD.

Tested on FreeBSD 8.1, CentOS 4.5, and Mac OS X 10.6.5