New default application template for create-app.
This contains a new template for creating new applications that
is more inline with what the users on the rebar mailing would
expect and find useful. It uses some of the simple template that
was there before, but instead puts the application in an 'apps'
folder. It also adds a rebar.config which points to this app,
and a makefile which includes a 'start' target to start the app
(as per tuncer's suggestion).

The new application will build and run cleanly.