New default application template for create-app
This contains a new template for creating new applications that is
inline with what users on the rebar mailing would prefer. Reuses parts
of simpletemplate. Contains rebar.config and a Makefile which has a
'start' target for running the application. This template builds and
runs cleanly from creation. rebar_templater updated to use this new
template by default.