Add 'upgrade' command
To support OTP release upgrades I have added support for building
upgrade packages. Support for this is included in the
rebar_upgrade module, specifically upgrade/2. It requires one
variable to be set on the command line 'oldreleasepath' which is
the absolute path or relative path from 'rel/' to the previous
release one is upgrading from. Running an upgrade will create the
needed files, including a relup and result in a tarball containing
the upgrade being written to 'rel/'. When done it cleans up the
temporary files systools created. Usage: "./rebar upgrade
oldreleasepath=/path/to/old/version" This also includes a dummy
application that can be used to test upgrades aswell as an example.