Add preprocess support for edoc
While trying to build the docs for riak_core I was bumping up against
a problem where edoc would fail because ?VNODE_REQ{...} is invalid
syntax.  By default, edoc does not run the preprocessor, but it can be
enabled via the {preprocess, true} option.  By doing this, the above
syntanx will get translated and edoc won't fail.  In order for this to
work all necessary include files need to be passed via the {includes,
[...]} option.

This patch will check for the preprocess option and, if found, will
build the list of include directories needed and pass it to edoc.