Support for setting deps_dir in $HOME/.rebar/config
This patch adds an additional check when rebar is setting the global
deps_dir, to see if it is defined in the global config file (i.e.,
the one in $HOME/.rebar). This makes life a lot easier when working
with multiple projects that use many of the same deps, as the location
you set in your global config can be used for each if not overriden by
the local config file(s).