Remove escript from code path before running eunit.
It appears that when the erlang loader finds the rebar escript on the code
path, which does _not_ have the module it is attempting to load it results in
the following error message being logged.

=ERROR REPORT==== 3-Aug-2011::09:21:21 ===
File operation error: bad_central_directory. Target:
/path/to/rebar/calendar.beam. Function: get_file. Process: code_server.

For a sizeable project with a comprehensive test suite this can introduce
quite a lot of noise into the rebar eunit output.

We can work around this problem by removing the rebar escript from the code
path just before we run the eunit tests.