Add root_dir option to reltool.config
When generating a node using rebar generate, rebar always used code:root_dir() to find the Erlang installation to clone into the node tree.  However, for anyone wishing to build a cross-compiled node, there was no way to override this.  This patch adds a new option to reltool.config file to allow an alternate root_dir to be specified, for example:

{root_dir, "/tmp/otp_R14B03_armv7l"}

This assumes that the contents of /tmp/otp_R14B03_armv7l has been generated using the instructions found in the xcomp directory in your $ERL_TOP structure.  NOTE: you may have to add additional filters to exclude files not present in the xcomp version, such as the .smp files if you have disabled that.