Allow reltool target_dir to be constructed on the fly
The version info in reltool.config can already be expanded by rebar.
This patch makes it possible to use release name and version info to
dynamically construct a target_dir. The patterns $(name) and $(vsn)
are automatically expanded to the corresponding values in RelToolConfig.

For example:

{target_dir, "$(name)/$(vsn)"}

would result in a target directory like "rel/mynode/2.2.1".