rebar should expand VCS version in the top directory, if possible
When a VCS tag is present in the reltool.config file, rebar tries
to expand it. However, it does so in the rel/ directory, which
(at least for git) may not reveal the latest version.

A problem is that if rebar was called with `cd rel; rebar generate`,
rebar will not know what the top directory is. This can be solved
by instead running `rebar generate skip_deps=true` from the top
directory. This patch makes rebar_rel_utils use base_dir instead
of the directory of reltool.config.