Provide rebar_utils:consult_files API and use it for rebar.config
This allows a developer to provide custom handlers for parsing
files other than rebar.config and rebar.config.script. The alternative
values must be added under rebar_config_formats as a list of two-item
tuples, the first item is the filename and the second is a tuple
containing a module and a function to invoke in case the file
is found. The invoked callback receives the file full path as
argument and must return { ok, Terms } if it succeeds,
{ error, Reason} otherwise.