use full paths where possible to be friendly to third party tools
This is a somewhat intrusive change that makes rebar use full paths in
most places. It should also do the same for many plugins that use
rebar_utils. The goal of this patch is to make the output of rebar
consumable by external applications like emacs. Basically, it makes the
directores output by rebar no longer dependent on where rebar is running.