Changing escriptized rebar permissions to a+x
By default, executables in paths such as /bin and /usr/local/bin
have the mode saying they can be executed by all.

The current version of rebar only has the user able to execute it,
which creates problems when copied directly in repositorie and
requiring other programs to interact with them.

This change makes rebar follow the standard of linuxes and unixes
by setting the permission flag to a+x, allowing users, the group
and others to execute it.