Stop cover between eunit runs for speed
Cover gets slower and slower for each application. This is due to
the cover_server internal state. Stopping the cover between eunit+cover
runs, emptying the cover_server state, gives a ~5-6 speed improvement
on systems with a lot of dependencies. Stopping the cover replaces the
earlier practice of doing a cover:reset before each run. On a project
consisting of 62 dependencies with a total of 1866 Erlang modules the
running time of rebar eunit decreased from ~20 minutes to ~3 minutes.