Added retry logic to error handling for shelled out commands
This is an attempt to work around recent github flakiness. This commit
adds the option {error_handler, {retry, Fun}} to the options
rebar_utils.sh/2 understands. The retry function is called with the
command and error, as rebar does with existing error handlers, and
a third argument representing the number of retries attempted.

This commit also modifies the logic git cloning deps to retry when
the clone fails. In this case, retry means deleting the possibly
borked clone, waiting for some interval, and then retrying up to
three times.