Fix parse_transform option in rebar.config
Currently this option doesn't work very well because
given parse_transform is applied globally to all modules,
even to parse transform module itself which is not possible
because the module is not yet compiled.

So we just ensure that local parse transformations are not applied
to themselves and probably also to erl_first_files since they
are explicitly stated as to be compiled before anything else.