Add report and verbose options for LFE compile
Recent builds of LFE have stopped reporting compile errors when compiled with
rebar. Examination of this issue revealed that this was due to recent
changes in default LFE compile options.

The way that LFE handles compile options has slightly changed; ``verbose`` and
``report`` are now provided as defaults only if no other options are passed; if
other options are passed, these do not get set. As a result, we have stopped
seeing compile error info when building with rebar.

This change brings back the previous level of reporting that LFE developers had
when compiling their projects with rebar.