Introduce REBAR_DEPS_PREFER_LIBS env variable to alter search behaviour
When REBAR_DEPS_PREFER_LIBS is set, dependencies with defined sources
are allowed to be searched for in system lib directory.

Under specific circumstances (i.e. in build environments without networking)
it is impossible to fetch deps locally. So, user needs a way to ask rebar to
search in system lib directory as well.

Signed-off-by: Matwey V. Kornilov <matwey.kornilov@gmail.com>